Commit part 1:
- Added Stripe Account Controller so users can create stripe accounts
- When user registers, automatically gives them a stripe id
- Can create a bank account in the front end with back end support via StripeAccountController
- Started working on a UserPaymentService, responsible for handling transaction between users

+++ Core/Entities/Listing.cs
@@ -7,7 +7,7 @@ namespace Core.Entities
         public int VenueId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public double Pay {  get; set; }
+        public decimal Pay {  get; set; }
         public Venue Venue { get; set; }
         public ICollection<ListingApplication> Applications { get; set; }
         public ICollection<ListingGenre> ListingGenres { get; set; }
